Refactor : 포맷팅

+++ app/src/main/java/com/kuit/conet/UI/GroupMain/SideBar.kt
@@ -98,7 +98,7 @@ class SideBar(
             }
 
             R.id.ll_sidebar_group_delete -> {
-                val deleteGroupDialog = DeleteGroupDialog(groupMainActivity, groupId.toInt())
+                val deleteGroupDialog = DeleteGroupDialog(groupMainActivity, groupId)
                 destoryFragment()
                 deleteGroupDialog.show(parentFragmentManager, DeleteGroupDialog.DIALOG_TAG)
             }
